+++ GameEngine/RectangleShape.h
@@ -7,25 +7,25 @@ namespace Moo
 	class RectangleShape : public Shape
 	{
 	public:
-		RectangleShape(int width, int height, int x, int y, XMFLOAT4 color);
+		RectangleShape(float width, float height, float x, float y, XMFLOAT4 color);
 		~RectangleShape();
-		void move(int x, int y);
-		void setX(int x);
-		void setY(int y);
-		int getX();
-		int getY();
-		int getWidth();
-		int getHeight();
+		void move(float x, float y);
+		void setX(float x);
+		void setY(float y);
+		float getX();
+		float getY();
+		float getWidth();
+		float getHeight();
 		void release();
 		void draw(ID3D11Device *dev, ID3D11DeviceContext *devcon);
 		ID3D11Buffer* const* getVertexBuffer() const;
 	private:
 		ID3D11Buffer* _vertexBuffer;
 		VERTEX _vertices;
-		int _width;
-		int _height;
-		int _x;
-		int _y;
+		float _width;
+		float _height;
+		float _x;
+		float _y;
 		XMFLOAT4 _color;
 	};
 }
